+++ packages/next-intl/src/middleware/utils.tsx
@@ -2,76 +2,20 @@ import {
   Locales,
   LocalePrefixConfigVerbose,
   DomainConfig,
-  Pathnames
+  Pathnames,
+  DomainsConfig
 } from '../routing/types';
 import {
   getLocalePrefix,
+  getSortedPathnames,
   matchesPathname,
+  normalizeTrailingSlash,
   prefixPathname,
   templateToRegex
 } from '../shared/utils';
 
 export function getFirstPathnameSegment(pathname: string) {
   return pathname.split('/')[1];
-}
-
-function isOptionalCatchAllSegment(pathname: string) {
-  return pathname.includes('[[...');
-}
-
-function isCatchAllSegment(pathname: string) {
-  return pathname.includes('[...');
-}
-
-function isDynamicSegment(pathname: string) {
-  return pathname.includes('[');
-}
-
-export function comparePathnamePairs(a: string, b: string): number {
-  const pathA = a.split('/');
-  const pathB = b.split('/');
-
-  const maxLength = Math.max(pathA.length, pathB.length);
-  for (let i = 0; i < maxLength; i++) {
-    const segmentA = pathA[i];
-    const segmentB = pathB[i];
-
-    // If one of the paths ends, prioritize the shorter path
-    if (!segmentA && segmentB) return -1;
-    if (segmentA && !segmentB) return 1;
-
-    // Prioritize static segments over dynamic segments
-    if (!isDynamicSegment(segmentA) && isDynamicSegment(segmentB)) return -1;
-    if (isDynamicSegment(segmentA) && !isDynamicSegment(segmentB)) return 1;
-
-    // Prioritize non-catch-all segments over catch-all segments
-    if (!isCatchAllSegment(segmentA) && isCatchAllSegment(segmentB)) return -1;
-    if (isCatchAllSegment(segmentA) && !isCatchAllSegment(segmentB)) return 1;
-
-    // Prioritize non-optional catch-all segments over optional catch-all segments
-    if (
-      !isOptionalCatchAllSegment(segmentA) &&
-      isOptionalCatchAllSegment(segmentB)
-    ) {
-      return -1;
-    }
-    if (
-      isOptionalCatchAllSegment(segmentA) &&
-      !isOptionalCatchAllSegment(segmentB)
-    ) {
-      return 1;
-    }
-
-    if (segmentA === segmentB) continue;
-  }
-
-  // Both pathnames are completely static
-  return 0;
-}
-
-export function getSortedPathnames(pathnames: Array<string>) {
-  const sortedPathnames = pathnames.sort(comparePathnamePairs);
-  return sortedPathnames;
 }
 
 export function getInternalTemplate<
@@ -112,7 +56,7 @@ export function getInternalTemplate<
   }
 
   // Try to find an internal pathname that matches (this can be the case
-  // if all localized pathnames are different from the internal pathnames).
+  // if all localized pathnames are different from the internal pathnames)
   for (const internalPathname of Object.keys(pathnames)) {
     if (matchesPathname(internalPathname, pathname)) {
       return [undefined, internalPathname];
@@ -136,6 +80,10 @@ export function formatTemplatePathname(
   }
 
   targetPathname += formatPathnameTemplate(targetTemplate, params);
+
+  // A pathname with an optional catchall like `/categories/[[...slug]]`
+  // should be normalized to `/categories` if the catchall is not present
+  // and no trailing slash is configured
   targetPathname = normalizeTrailingSlash(targetPathname);
 
   return targetPathname;
@@ -231,12 +179,17 @@ export function getPathnameMatch<AppLocales extends Locales>(
 }
 
 export function getRouteParams(template: string, pathname: string) {
-  const regex = templateToRegex(template);
-  const match = regex.exec(pathname);
+  const normalizedPathname = normalizeTrailingSlash(pathname);
+  const normalizedTemplate = normalizeTrailingSlash(template);
+
+  const regex = templateToRegex(normalizedTemplate);
+  const match = regex.exec(normalizedPathname);
   if (!match) return undefined;
   const params: Record<string, string> = {};
   for (let i = 1; i < match.length; i++) {
-    const key = template.match(/\[([^\]]+)\]/g)?.[i - 1].replace(/[[\]]/g, '');
+    const key = normalizedTemplate
+      .match(/\[([^\]]+)\]/g)
+      ?.[i - 1].replace(/[[\]]/g, '');
     if (key) params[key] = match[i];
   }
   return params;
@@ -296,7 +249,7 @@ export function isLocaleSupportedOnDomain<AppLocales extends Locales>(
 export function getBestMatchingDomain<AppLocales extends Locales>(
   curHostDomain: DomainConfig<AppLocales> | undefined,
   locale: string,
-  domainConfigs: Array<DomainConfig<AppLocales>>
+  domainsConfig: DomainsConfig<AppLocales>
 ) {
   let domainConfig;
 
@@ -307,12 +260,12 @@ export function getBestMatchingDomain<AppLocales extends Locales>(
 
   // Prio 2: Use alternative domain with matching default locale
   if (!domainConfig) {
-    domainConfig = domainConfigs.find((cur) => cur.defaultLocale === locale);
+    domainConfig = domainsConfig.find((cur) => cur.defaultLocale === locale);
   }
 
   // Prio 3: Use alternative domain with restricted matching locale
   if (!domainConfig) {
-    domainConfig = domainConfigs.find(
+    domainConfig = domainsConfig.find(
       (cur) => cur.locales != null && cur.locales.includes(locale)
     );
   }
@@ -324,7 +277,7 @@ export function getBestMatchingDomain<AppLocales extends Locales>(
 
   // Prio 5: Use alternative domain that supports all locales
   if (!domainConfig) {
-    domainConfig = domainConfigs.find((cur) => !cur.locales);
+    domainConfig = domainsConfig.find((cur) => !cur.locales);
   }
 
   return domainConfig;
@@ -334,15 +287,15 @@ export function applyBasePath(pathname: string, basePath: string) {
   return normalizeTrailingSlash(basePath + pathname);
 }
 
-export function normalizeTrailingSlash(pathname: string) {
-  if (pathname !== '/' && pathname.endsWith('/')) {
-    pathname = pathname.slice(0, -1);
-  }
-  return pathname;
-}
-
 export function getLocaleAsPrefix<AppLocales extends Locales>(
   locale: AppLocales[number]
 ) {
   return `/${locale}`;
+}
+
+export function sanitizePathname(pathname: string) {
+  // Sanitize malicious URIs, e.g.:
+  // '/en/\\example.org → /en/%5C%5Cexample.org'
+  // '/en////example.org → /en/example.org'
+  return pathname.replace(/\\/g, '%5C').replace(/\/+/g, '/');
 }