+++ packages/next-intl/src/navigation/shared/redirects.tsx
@@ -5,7 +5,7 @@ import {
 import {Locales, LocalePrefixConfigVerbose} from '../../routing/types';
 import {ParametersExceptFirst} from '../../shared/types';
 import {getLocalePrefix, prefixPathname} from '../../shared/utils';
-import {isLocalHref} from './utils';
+import {isLocalizableHref} from './utils';
 
 function createRedirectFn(redirectFn: typeof nextRedirect) {
   return function baseRedirect<AppLocales extends Locales>(
@@ -18,7 +18,8 @@ function createRedirectFn(redirectFn: typeof nextRedirect) {
   ) {
     const prefix = getLocalePrefix(params.locale, params.localePrefix);
     const localizedPathname =
-      params.localePrefix.mode === 'never' || !isLocalHref(params.pathname)
+      params.localePrefix.mode === 'never' ||
+      !isLocalizableHref(params.pathname)
         ? params.pathname
         : prefixPathname(prefix, params.pathname);
     return redirectFn(localizedPathname, ...args);