+++ packages/use-intl/.size-limit.ts
@@ -4,19 +4,19 @@ const config: SizeLimitConfig = [
   {
     name: './ (ESM)',
     path: 'dist/esm/index.mjs',
-    limit: '15.25 kB'
+    limit: '13.405 kB'
   },
   {
     name: './ (no useTranslations, ESM)',
     path: 'dist/esm/index.mjs',
     import:
-      '{useLocale, IntlProvider, useNow, useTimeZone, useMessages, useFormatter}',
+      '{IntlProvider, useLocale, useNow, useTimeZone, useMessages, useFormatter}',
     limit: '1.825 kB'
   },
   {
     name: './ (CJS)',
     path: 'dist/cjs/index.js',
-    limit: '19.775 kB'
+    limit: '17.475 kB'
   }
 ];